Update 04_01_download_length.py
Fix a bug for Python 3.6 using Anaconda

+++ 04_01_download_length.py
@@ -1,6 +1,6 @@
 BUFF_SIZE = 1024
 def download_length(response, output, length):
-    times = length / BUFF_SIZE
+    times = int(length / BUFF_SIZE)
     if length % BUFF_SIZE > 0:
         times += 1
     for time in range(times):
